+++ components/keyscreen/emulate/src/main/java/com/flipperdevices/keyscreen/emulate/composable/common/ComposableEmulateButtonWithText.kt
@@ -15,6 +15,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.theme.LocalTypography
 import com.flipperdevices.keyscreen.emulate.composable.common.button.ComposableEmulateButton
@@ -47,7 +48,7 @@ fun ComposableEmulateButtonWithText(
         )
 
         Row(
-            modifier = Modifier.padding(top = 4.dp),
+            modifier = Modifier.padding(top = 0.dp),
             verticalAlignment = Alignment.CenterVertically
         ) {
             if (iconId != null) {
@@ -62,7 +63,7 @@ fun ComposableEmulateButtonWithText(
             }
             Text(
                 text = textId?.let { stringResource(it) } ?: "",
-                style = LocalTypography.current.subtitleM12,
+                style = LocalTypography.current.bodyR14.copy(fontSize = 12.sp),
                 color = LocalPallet.current.text30
             )
         }